requirements.txt
web3==6.15.1
eth-account==0.10.0
requests==2.31.0
python-dotenv==1.0.0
colorama==0.4.6
tabulate==0.9.0
tqdm==4.66.1
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

